added tokenExtractor middleware

+++ controllers/blogs.js
@@ -1,23 +1,29 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
-blogsRouter.use('/', (req, res, next) => {
-  next()
-})
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('author', { username: 1, name: 1 })
   response.status(200).json(blogs)
 })
+
 blogsRouter.get('/:id', async (request, response) => {
   const { id } = request.params
   const blogToGet = await Blog.findById(id)
   response.status(200).json(blogToGet)
 })
+
 blogsRouter.post('/', async (request, response) => {
   const { body } = request
-  const user = await User.findById(body.userId)
+  const { token } = request
+  const decodedToken = await jwt.verify(token, process.env.SECRET)
+  console.log('decodedToken :>> ', decodedToken)
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token is missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   const blogToAdd = new Blog({
     title: body.title,
@@ -28,13 +34,24 @@ blogsRouter.post('/', async (request, response) => {
   const savedBlog = await blogToAdd.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
-  response.status(201).json(savedBlog)
+  return response.status(201).json(savedBlog)
 })
+
 blogsRouter.delete('/:id', async (request, response) => {
   const { id } = request.params
+  const { token } = request
+  const decodedToken = await jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token is missing or invalid' })
+  }
+
   await Blog.findByIdAndRemove(id)
-  response.status(204).end()
+  const user = await User.findOne({ username: decodedToken.username })
+  user.blogs = user.blogs.filter((blog) => blog.id !== id)
+  await user.save()
+  return response.status(204).end()
 })
+
 blogsRouter.put('/:id', async (request, response) => {
   const { id } = request.params
   const { body } = request
